requirements.txt
pytelegrambotapi>=4.15.0
openpyxl>=3.1.2
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.2